moved utils to app
