+++ app/util/checkFields.js
